+++ libs/qt/source/ftrack_qt/widgets/dialogs/__init__.py
@@ -5,3 +5,4 @@ from ftrack_qt.widgets.dialogs.scroll_tool_configs_dialog import (
     ScrollToolConfigsDialog,
 )
 from ftrack_qt.widgets.dialogs.tab_dialog import TabDialog
+from ftrack_qt.widgets.dialogs.radio_button_dialog import RadioButtonDialog