+++ libs/qt/release_notes.md
@@ -2,6 +2,11 @@
 
 ## upcoming
 
+* [new] MenuButton; Button supporting menu if multiple action items provided.
+* [changed] FileDialog; support selecting multiple files.
+* [changed] FileBrowser; support collection of multiple files.
+* [changed] Python clique 1.6.1 library as dependency.
+* [new] RadioButtonDialog; Dialog to return an item from a list of items represented by radio buttons.
 * [changed] Improved AccordionWidget and AcordionHeaderWidget to use EditableLabelWidget and possibility to add a remove button. Emits bin_clicked, title_changed, title_edited signals.
 * [fix] BuildProgressData utility, convert options to string.
 * [changed] StatusMaterialIconWidget clickable.