+++ libs/qt/release_notes.md
@@ -1,6 +1,18 @@
 # ftrack QT library release Notes
 
+
 ## upcoming
+
+* [changed] FileDialog; support selecting multiple files.
+* [changed] FileBrowser; support collection of multiple files.
+* [changed] Python clique 1.6.1 library as dependency.
+* [new] RadioButtonDialog; Dialog to return an item from a list of items represented by radio buttons.
+* [change] build_progress_data; Check enabled/disabled plugins.
+* [change] AccordionBaseWidget; Emit enabled event signal.
+
+
+## v2.2.2
+2024-06-26
 
 * [fix] Fix invoke_in_qt_main_thread so it properly returns the result of the function.